requirements.txt
requests==2.25.1
pyyaml==5.4.1
tabulate==0.8.9
cassandra-driver==3.25.0
pytest==6.2.3
coverage==5.5